+++ src/navigation/dashboardActions.ts
@@ -2,12 +2,14 @@ import { DashboardAction } from './types';
 
 import { dashboardEntry as aboutEntry } from '../screens/AboutScreen';
 import { dashboardEntry as exportKeyEntry } from '../screens/ExportKeyScreen';
-import { dashboardEntry as keycardMenu } from '../screens/KeycardMenuScreen';
-import { dashboardEntry as addressMenu } from '../screens/address/AddressMenuScreen';
+import { dashboardEntry as keycardMenuEntry } from '../screens/KeycardMenuScreen';
+import { dashboardEntry as settingsEntry } from '../screens/SettingsScreen';
+import { dashboardEntry as addressMenuEntry } from '../screens/address/AddressMenuScreen';
 
 export const dashboardActions: DashboardAction[] = [
   exportKeyEntry,
-  addressMenu,
-  keycardMenu,
+  addressMenuEntry,
+  keycardMenuEntry,
+  settingsEntry,
   aboutEntry,
 ];